Traktor: Static code analysiz fixes. VS 2015 project files.
[git-p4: depot-paths = "//depot/traktor/": change = 17956]

+++ code/Core/Io/BufferedStream.cpp
@@ -165,7 +165,7 @@ int BufferedStream::read(void* block, int nbytes)
 	uint8_t* out = static_cast< uint8_t* >(block);
 	uint8_t* end = out + nbytes;
 
-	if (nbytes <= m_internalBufferSize)
+	if (nbytes <= int32_t(m_internalBufferSize))
 	{
 		// Read and copy until number of desired bytes read is meet.
 		while (out < end)
